payment-service/requirements.txt
dapr==1.13.0
dapr-ext-workflow>=0.1.0
fastapi==0.109.1
grpcio==1.62.0
pydantic==2.4.2
requests==2.31.0
uvicorn==0.23.2
email-validator==2.1.1
stripe==9.9.0